Move VPDQIndexMatch to VPDQIndex
