+++ python-threatexchange/threatexchange/extensions/vpdq/vpdq_index.py
@@ -9,7 +9,6 @@ import typing as t
 import vpdq
 from threatexchange.signal_type.index import (
     SignalTypeIndex,
-    VPDQIndexMatch,
     IndexMatch,
     T as IndexT,
 )
@@ -20,6 +19,17 @@ from threatexchange.extensions.vpdq.vpdq_util import (
     VPDQ_DISTANCE_THRESHOLD,
 )
 
+class VPDQIndexMatch(IndexMatch[IndexT]):
+    def __init__(
+        self,
+        distance: int,
+        query_match_percent: float,
+        compared_match_percent: float,
+        metadata: T,
+    ) -> None:
+        super().__init__(distance, metadata)
+        self.query_match_percent = query_match_percent
+        self.compared_match_percent = compared_match_percent
 
 class VPDQIndex(SignalTypeIndex[IndexT]):
     """
